+++ src/common/utxobased/info/bitcoin.ts
@@ -53,7 +53,7 @@ const currencyInfo: EdgeCurrencyInfo = {
 const engineInfo: EngineInfo = {
   serverConfigs: [
     {
-      type: 'blockbook-nownode',
+      type: 'blockbook',
       uris: ['https://btc-wusa1.edge.app', 'https://btc-eu1.edge.app']
     },
     {