+++ src/common/utxobased/info/pivx.ts
@@ -44,7 +44,7 @@ const currencyInfo: EdgeCurrencyInfo = {
 const engineInfo: EngineInfo = {
   serverConfigs: [
     {
-      type: 'blockbook-nownode',
+      type: 'blockbook',
       uris: ['https://pivx-wusa1.edge.app']
     },
     {